Get the popup dialogs to work right
Now, everything works and the newly downloaded update is loaded, but it fails
because only checked in files are downloaded. This effectively means that we
need separate deployment branches where everything is checked in, since we
don't really want to check everything into the dev branches.

I don't like this new git-based workflow that much. It doesn't seem to work
well for ionic1 builds since the build step is skipped. Maybe it works better
for ionic2 since the build step is actually executed and could pull the
standard libraries.