Manually add the automatically installed libraries to this branch
We can then use this as the base for the branches that we want to deploy to
ionic.

So in order to deploy a new channel, we:
- checkout a new, temporary branch from this one
- merge the branch to deploy into the temp branch
- commit and push to ionic
- delete the temp branch

These changes (to add the new libraries installed from bower) should never be
merged into master. Also, if we have a separate template for deployable code,
we can delete the habitica avatar stuff from it instead of from each of the
branches, which will make the branches much easier to use because they won't
have this big honking diff from master.