register for listeners *after* init
This looks a bit weird because the init generates the registration message, but
if the listener is not registered in time, it will miss the message. HOWEVER,
if we try to move it earlier, which is what I first did, the object has not yet
been created, so you get null. I don't see a different way of using the API -
init creates the object, you need to listen to notifications on the object and
init also (in the background) generates the notifications.

So let's go with this for now and hope that the timing is right.

I am not sure how this was not caught it testing.