Save the new client tag to the user's profile
Now that we are beginning to run into the use case with multiple clients on the
same server (definitely emotion v/s information, and potentially SDSU), need to
store the client on the server when the user switches to it.

If we assume that users will potentially switch between multiple clients with
different policies, we need to store the switch as a message. But we don't have
a use case for that, and it is harder to work with, so let's go with the cheap
solution for now.