Ensure that the L+ version has the biggest code
Consistent with the multi-APK documentation, change the android build so that
the L+ version is base+9, while the native pre-L versions are base+2 and
base+4.

This also means that we can do native versions of pre-L and don't have to
remove the game functionality. We need to check in a new gradle build file to avoid the flavors error.
https://stackoverflow.com/a/47418102/4040267