Handle more native usercache <-> local storage mismatch
We ran into another use case in which there was a mismatch between the native
usercache and the local storage, this time on android. Unfortunately, due to
lack of appropriate error messages to the user, we didn't know when this
happened, and so the root cause is lost.

But we already fix this for the consent case - we can also fix it for any
dual-stored entries.

This fixes #338. We should really improve the user-visible error logging on
sync failures so that we can catch this when we have a chance to root cause
it.