Add an end trip + force sync button
To force the push of other entries (e.g. mode confirmation, incidents)
entered after a trip is complete. Normally, these would be pushed when the
next trip ends, but for development and debugging, it is useful to
be able to force the push of all entries.

Normally, we would do this by manually starting and ending a trip, but tis is
faster and does not rely on networking set up correctly on the emulator. The
android emulator, for example, has device-only networking turned on by
default.