Format changes to the index.html so that the phonegap rewrite works
Phonegap rewrites index.html to add support for live reload. The new code
(injected from https://github.com/phonegap/connect-phonegap) open a websocket
to the `phonegap serve` server. The websocket sends notifications of javascript
changes so that the client can update.

But this assumes that the current CSP is in one line, and cordova.js is tagged
with javascript. Let's make the appropriate changes so that live reload works
for our app as well.