Handle weird behavior of files loaded from the library
After the code is downloaded using `$ phonegap serve`, it is loaded from a
different path than usual. While accessing files from this path using `$http`,
we get a blank string if the file does not exist. In the default app, we get an
error callback, as verified by installing it and re-running.

Handle this by checking for empty string and throwing an exception in that case.