Make native calls only after the platform has initialized
Since the JS code is reloaded, it frequently executes before the platform is
re-initalized. This can happen even in the regular client, but happens almost
all the time on reload, and when it happens, everything breaks because the
native methods are undefined.

To fix, we put the native startup methods into a `$ionicPlatform.ready()` so
that they are not executed until the native platform is ready.

Only a couple of other changes needed for `$ ionic serve` to work -
a) Move all the habitica code out - other branches are not using it, and
including it by default makes things Very Slow.
b) copy over the plugin-specific javascript files

```
www/js/control/collect-settings.js
www/js/control/sync-settings.js
www/js/control/transition-notify-settings.js
```