Split out the habitica classes into their own repository
The new repository is
https://github.com/e-mission/e-mission-habitica-local

This allows us to use live reload to test the phone code with the devapp

Without this change, the code that needs to be downloaded is so large that the
download appears to hang.
https://github.com/e-mission/e-mission-phone/pull/412/commits/c40fb7422195d1aa235e6feb1b7160acc55cf45e

While we could continue deleting this directory from each individual branch, as
we have been doing, none of the current projects plan to use habitica :(
And this can then be the first of the changes to break up the phone UI and
convert it from being monolithic into a set of plugins.