Switch to primarily native storage
Back in the old days, we tried to move all storage to local storage so that we
could read it without waiting for ionicPlatform.ready and startup quickly.
Well, it turned out that sometimes local storage was deleted on iOS, so we
ended up with the case where local intro_done was deleted, so we would reset the UI.
Which looked stupid and like a bug (https://github.com/e-mission/e-mission-phone/issues/443).

Re-writing to pull out the KV storage into a service, write primarily to native
with local as backup, and read from both and unify.

Everything was already promisified because we were reading consent from native
code, so I am not sure that this actually affects performance. And it is a heck
of a lot easier to understand.

Testing done:
- started app -> went to intro
- restarted app after restarting emulator -> went to current screen. For fixes to the current screen, see 0ecfdaecce66b2bd9be3fab6efe386eeaab26287
- cleared out intro_done and consent from local storage and restarted app -> went to current screen
- cleared all native storage, generated popups for intro_done and consent, then crashed due to https://github.com/e-mission/e-mission-transition-notify/issues/18 On relaunch, launched properly without any popups. I declare that this is done.

```
[phonegap] [console.log] DEBUG:uc_stored_val = null ls_stored_val = {"intro_done":"2018-09-25T13:14:51-07:00"}
```