Remove unused `angularLocalStorage` module
We want to move away from angularLocalStorage, since local storage can be
deleted by the OS, at least on iOS, when the device is low on disk space.
This means that we can't store persistent data in local storage, only session
data.

In a09385b03cbcc994c0bd3c33d7317a242132dc65, fixed this for the startup code.
This is the first of a series of changes that will refactor other code code to
remove all instances.

In this change, we remove the module dependency if it is no longer used.
Simple changes!