Change over the goals/referrals code to the new storage
goals code in general switches to native storage, but
the referral code alone, since it is not intended for long-term storage, uses
the new local-only storage.

Note that we can change the referral code to store a single object that
includes all referral objects, e.g.

{
    REFERRED_KEY:
    REFERRED_GROUP_ID:
    REFERRED_USER_ID:
    REFERRAL_NAVIGATION_KEY:
}
...

I had initially implemented this change, but the navigation key is modified
separately from the others, so we would need to deal with it correctly, and I
don't really want to re-test this code when nobody is actively using it.

Bonus fix: clean up the survey code that we had marked with a TODO