Convert the incident range also into a promisified native storage
This was a bit tricky because of the way that the slider is hooked into the
variable.

If you read the data immediately without promisifying AND the load happens
before the store, then the slider is set to the loaded value and the store
retains the correct value.

HOWEVER, if you read the data through a promise, then you get the store first,
and without additional checks, the value is reset to the default.

The solution is to add additional checks to only store valid values.