Remove the code that sets connection and sync default settings
Having the connection settings, in particular, be included is a challenge. If
trip state notifications are used, then this cause the app to crash
consistently on re-launch.

This should fix
https://github.com/e-mission/cordova-jwt-auth/issues/33

The tripaware studies did not include this code - e.g.
https://github.com/e-mission/e-mission-phone/blob/urap-2017-control/www/js/app.js
https://github.com/e-mission/e-mission-phone/blob/urap-2017-emotion/www/js/app.js
https://github.com/e-mission/e-mission-phone/blob/urap-2017-information/www/js/app.js
and worked just fine, so it is good that we discovered this early in this cycle.