Import fixes to the plugin javascript
The plugin javascript is not checked in to `www/js` for obvious reasons.
But we need it in the www directory that is downloaded for ionic.
So we copy it over and check it into this new branch.
As we make changes to the plugin javascript, we need to reflect it here.