Handle both pushed and unpushed mode confirmations for local and remote data
This was actually **harder** to do while keeping the local code from @sunil07t
in place, but I did it anyway to keep the changes simple. I am definitely going
to rip them out in the next commit.

This was actually fairly complex because of all the retained
getLocalModeForTrip and getLocalModeForPurpose, and its inconsistency with the
data in the timeline. In particular, the data in the HTML is tripgj objects
which have their start_ts in trip_gj.data.properties.start_ts. But
getLocalModeForTrip is called with the trip object, which has the start_ts in
trip.properties.start_ts. I had to unify that by passing assuming that the
passed in trip object had the start_ts field in it, and then passing in
trip.data.properties or trip.properties as the case may be.

Similarly, modes retrieved remotely using the unified data loader have the full
{'data': ..., 'metadata': ...} format but we were retrieving local data without
metadata. Unified that by retrieving local data with metadata, but then we now
need to set internal variables to mode.data instead of mode so that the labels
can be read from them.

Testing done:
https://github.com/e-mission/e-mission-docs/issues/353#issuecomment-481966262
https://github.com/e-mission/e-mission-docs/issues/353#issuecomment-481967505