Pull out the list of modes and purposes into a separate json file
Note that there is still a `.sample` file checked in, so everything will
continue to work out of the box, but people will not need to modify code
to change the set of modes and purposes any more :)

+ add new service for common functionality (similar to incident)

Should we move the other choice also to the common services file?
Unsure. There is some difference between the two implementations, and really
the only common part is popping up a popup for the other mode.

Testing done:
- Displayed existing selected mode and purpose
- Changed mode and purpose from the list view
- Set mode and purpose from post-trip prompt and confirmed that they were
  visible when viewed in the list view