Make "other" mode even better
To avoid repeats of embarrassing errors such as
8ea787b5dcffa98858a49e7e2096fd0ba5dff40d

In particular, if the mode/purpose are not in the list of options, create a new
entry for them and add it to the valid list.

This has many benefits:
- we can now set the value to the selected mode/purpose
- if this is a common mode for the user, they can select the mode/purpose again

Unfortunately, this only work on a day by day basis. We should really consider
storing the user-specific other mode list locally and putting it into the
options every time.