Fix the spacing in the list view to not partially hide the mode and purpose buttons
So the reason that the spacing was incorrect is because the draft was displayed
below the mode as a new column, which moved everything down and caused the
buttons to go off the edge.

Fixed this by replacing the mode column (? and 100%) with a button that says
DRAFT and explaining what that means on click. This takes up a lot less room,
and fixes the spacing.

TESTING DONE:
- Viewed a draft mode trip - works
- Viewed an analysed trip - works

See screenshots in the associated PR