Implement initial integration with enketo-core to display complex embedded surveys
The form is displayed inline in the main app, so it can be manipulated much
more easily. In particular:
- both the form and the model are javascript objects passed in to a function so
  can be manipulated at will to customize the survey for a particular user.
- The result is also retrieved in the form, so results can be reflected
  immediately in the UX

The form is currently displayed as a view in the metrics tab, with its own
state.  I considered making it a popover instead, but wasn't sure that the
smaller width would work for the survey. I might experiment with that in a
future change.

The main changes to the app itself are to the main.js to define the new state,
and a new survey launcher `enketo_launch.js` that is the controller for the new
state.

The rest of the code is to check in the enketo javascript libraries. It is
theoretically possible to install them via `npm install` and then use `rollup`
or `webpack` to create the bundle. But we don't really have a gruntfile or
other build mechanism to invoke the rollup/webpack, and ES6 imports don't
necessarily play nice with angular 1. For example, when I tried using `import {
Form } from `enketo-core`, angular complained about the import.

Also, the resulting bundle file did not include an export for `Form` since
rollup inlines functions where possible. I had to manually add
`window.FormModule` to the javascript file to allow it to be used from the rest
of the app.

Given all this, I decided to check in enketo-core as a `manual_lib`.

This change can be tested by applying the following patch, although it is
pretty kludgy. Long-term, individual projects need to determine where they want
to launch the test from.

Also includes a couple of sample JSON files for testing.

The input to the library is the enketo paper formatted JSON files.

They can be obtained using the following process:
- Create the form using kobotoolbox https://github.com/kobotoolbox/ or https://build.opendatakit.org/ This will let you download an xlsform
- Convert xlsform -> xml form using http://xlsform.opendatakit.org/
- Convert xml form -> enketo paper format form + model using
  enketo-transformer. The curl option works fine - I used it with requests
- Copy the json files into www/json
- Launch them by passing the JSON file as a parameter while changing state, as in `metrics.js` below

```
diff --git a/www/js/metrics.js b/www/js/metrics.js
index d80e81d7..e7e4a3d3 100644
--- a/www/js/metrics.js
+++ b/www/js/metrics.js
@@ -35,6 +35,13 @@ angular.module('emission.main.metrics',['nvd3', 'emission.services', 'ionic-date
         $scope.onCurrentTrip();
     });

+  $scope.testSurvey = function() {
+    $state.go("root.main.enketosurvey", {
+        form_location: "json/enketo-survey-sample-full.json"
+    });
+  }
+
+
     // If we want to share this function (see the pun?) between the control screen and the dashboard, we need to put it into a service/factory.
     // But it is not clear to me why it needs to be in the profile screen...
     var prepopulateMessage = {
diff --git a/www/templates/main-metrics.html b/www/templates/main-metrics.html
index 124e95d9..923e0d05 100644
--- a/www/templates/main-metrics.html
+++ b/www/templates/main-metrics.html
@@ -10,6 +10,9 @@
       <button class="button button-icon" ng-click="doRefresh()">
         <i class="icon ion-refresh"></i>
       </button>
+      <button class="button button-icon" ng-click="testSurvey()">
+        TEST SURVEY
+      </button>
   </ion-nav-buttons>
   <ion-content class="has-header">
     <div style="height: 40px; width: 100%; padding: 10px;" ng-if="uictrl.showVis"> <!-- visualization control -->
```