Fix the survey to support a wider variety of question types
This required some deep digging into the bowels of both enketo and ionic code,
but it works now. Please see
https://github.com/e-mission/e-mission-docs/issues/376 for way more detail than
can fit here :)

At this point, the survey is displayed in the e-mission app in the same way as
in a mobile browser, but returns values to the app on validation.