Convert almost all functions from the listview to use variables instead
Given the overhead of the functions, and the fact that they are called
multiple times as the view is refreshed, this is likely to improve performance.
There are still some minor function calls (e.g. to get the modes and colors)
and I should really investigate why we are mapping the incidents over and over.
But this is a good first step.