Rename all `displayName` to be consistent with `display_name`
For reasons that are now lost to the mists of time, we use the `displayName`
property to display the name. But the server fills it in as `display_name`
https://github.com/e-mission/e-mission-server/blob/master/emission/analysis/intake/cleaning/clean_and_resample.py

And although multiple people have modified the diary so far, nobody has
actually changed this. So in spite of the fact that we look up the place name
on the server, we look up the place name again on the phone, this time using
the OSM nominatim. Which is just wasteful and inefficient and not very kind to OSM.

Fixing now...

Bonus fix: only check the trip state if the platform is ready, since it
involves a native call. This ensure that the page reloads successfully. This is
a follow-on to df0166141e9357f8326ad1bbdd5f16c64c775a89 which restored the current trip functionality