Improve the diary screen to also use variables instead of functions
We re-use the formatted data from the trip as much as possible, but leave the
formatting function calls to the section data unchanged. We can change this too
if required, but it is unclear if it is required, since we only have a few
sections (max 5?) per trip, and they largely fit on one page without a lot of
scrolling.

In order to make this happen, I also had to create a tripWrapper map in the
Timeline so that the detail screen could look up the trip wrapper for the
formatting directly.

We should really rethink trip wrapper storage, since some of the items there
(e.g. the classes are really for display, but some of them (e.g. mode
confirmations) are really part of the associated data). But the modes interact
with the list of valid modes in the list.js, so it is non trivial to move it
out without another significant refactoring. which I don't plan to undertake now.

Finally, I fix two small issues that have annoyed me since forever.
- First, if the display name has already been filled in on the server side, we
  don't query OSM's nominatim again
- Second, the detail view also now displays the local time of the **trip**
  instead of the current local time. So for example, if I went to Hawaii and
  left on a trip at 10am Hawaii time, and then tried to review my trip diary
  after coming home to California, the trip should still say that it started at
  10am, even though that is 1pm PDT. @krisma fixed this in the list view but not
  the detail view. Moved this to the DiaryHelper and then called it in the
  detail view as well (getFormattedTime -> getLocalTimeString) all throughout.
  This fixes https://github.com/e-mission/e-mission-docs/issues/145