Defer reading the list of points until the button is actually clicked
While scrolling with the list, I saw several messages of the form

```
[Log] DEBUG:section 2016-08-04T13:10:38.739684-10:00 -> 2016-08-04T13:20:44-10:00 bound incident addition  (index.html, line 125)
[Log] DEBUG:Called getSection points with list of size 22 (index.html, line 125)
[Log] DEBUG:About to map point[-155.9108361,19.4228153] at index 0 (index.html, line 125)
[Log] DEBUG:Mapped point [-155.9108361,19.4228153] to [object Object] (index.html, line 125)
[Log] DEBUG:points 4:10 PM -> 4:20 PM bound incident addition  (index.html, line 125)
...
[Log] DEBUG:section 2016-08-04T10:03:51.235000-10:00 -> 2016-08-04T10:35:12-10:00 bound incident addition  (index.html, line 125)
[Log] DEBUG:Called getSection points with list of size 64 (index.html, line 125)
[Log] DEBUG:About to map point[-154.9029399,19.5461465] at index 0 (index.html, line 125)
[Log] DEBUG:Mapped point [-154.9029399,19.5461465] to [object Object] (index.html, line 125)
[Log] DEBUG:points 1:03 PM -> 1:35 PM bound incident addition  (index.html, line 125)
```

I had not added any incidents, and it is in fact not possible to add incidents from the list view. So why was I mapping lists of points with sizes 22, 270, 64, ...
While all these points are in memory, we still need to iterate over them and
map them, which is non-trivial. So I looked into where we were calling this from.

The answer is from the `onEachFeature` function of the diary helper, which
registers the incident adding code as a `click` handler for the trip geojson
directive. We do this since we re-use the same directive in both the list and
detail views, and we add incidents from the detail view map.

However, the incident handling code has multiple entry points, and two of them
deferred to the third, which returned a curried function that implicitly
referenced the data points. But that meant that the data points had to be read
and mapped when the function was *registered* not when it was *invoked*.

And as the list scrolls, angular cleans up and recreates directives, so this
happened over and over again.

Fixed by pulling out a non-curried implementation, and then returning curried
functions from every entry point, that read the points and then passed it to
the implementation function.

After the fix, the logs are:

```
[Log] DEBUG:section 2016-08-04T14:06:52.592000-10:00 -> 2016-08-04T14:11:33.160000-10:00 bound incident addition  (index.html, line 125)
[Log] DEBUG:section 2016-08-04T14:18:35.840464-10:00 -> 2016-08-04T14:34:12.571000-10:00 bound incident addition  (index.html, line 125)
index.html:125DEBUG:section 2016-08-04T14:39:38.288795-10:00 -> 2016-08-04T16:18:54.709000-10:00 bound incident addition
index.html:125DEBUG:section 2016-08-04T16:34:45.744782-10:00 -> 2016-08-04T16:38:38.348000-10:00 bound incident addition
index.html:125DEBUG:section 2016-08-04T13:10:38.739684-10:00 -> 2016-08-04T13:20:44-10:00 bound incident addition
index.html:125DEBUG:section 2016-08-04T10:41:32.136385-10:00 -> 2016-08-04T12:55:48.721000-10:00 bound incident addition
index.html:125DEBUG:section 2016-08-04T10:03:51.235000-10:00 -> 2016-08-04T10:35:12-10:00 bound incident addition
```

So the points are not read during registration and are instead read when the detail map is clicked.

```
DEBUG:points 1:03 PM -> 1:35 PM received click event, adding stress popup at LatLng(19.50527, -154.95512)
```