Fixed iOS offset issue
It turned out to be due to the enketo paging module scrolling to show the page
element that needs the focus. Since we are using ionic scrolling, it looks like
it already scrolls to show the top element, which works.

Please see the next 5-6 comments after
https://github.com/e-mission/e-mission-phone/pull/563#issuecomment-492356636

This doesn't _seem_ to break anything, but let me know if it does!