Principled implementation for the destination business id
Only missing "other" functionality.

Main changes:
- candidate businesses are now retrieved as part of the timeline, just like all
  remote data
- as part of `populateModeFromTimeline`, the candidate businesses are mapped to:
    - the popup JSON object
    - the value2entry and text2entry maps
Note that all of these are now `tripgj` specific instead of `$scope` specific

If the current business id is not in the candidate list, we retrieve its
details to get the name. Note that this is currently untested since we don't
support "other" functionality.

Note also that I have intentionally left all the `mode` variable names
untouched, so although the popup works and the selected choice is stored and
retrieved and displayed, it is stored as a `mode_confirm` object so the server
side code won't use it.

Finally, when the popup is opened, map the selected trip's options, and
xxx2entryMaps to `$scope`. This is required because we can only pass in a
`$scope` to the popup to retrieve the values to display. I tried passing in
`tripgj` instead of `$scope` and it failed because it couldn't call `$new` on
it. Fortunately, we create a temporary `$scope.draftMode` in `openModePopup` as
well, so I just have to unset the options and map in the same places where it
is unset (set to `angular.undefined`)

Testing done:
- Clicked on "mode" button
- Popup shows candidate destinations
- When one is selected, it is displayed
- The displayed value is retained even when the app is restarted

Please see video in related PR