Enable display of the settings for the current evaluation trip
Using a modal (as we did before) does not zoom in the trip details for reasons
that I don't understand. However, embedding the map in the current view does.
So we make the trip card also expandable, put the map and start/stop buttons
into the trip card, and expand it by default when a trip is selected.

Also implement the start/stop buttons to set some internal state and get
enabled/disabled properly.

Also ensure that the calibration card is hidden when the evaluation is selected.
Also add a refresh button that is currently unused.

Bonus fix: convert some variables to constants