Actually hook up the experiments
Along with some modifications based on further observations.

1. We really need to use the same techniques for calibration and evaluation
settings. Either the settings are applied when the new regime is selected, OR
they are applied later through a button. But we should use the same approach
for both. Because in the case of evaluation, we want to evaluate at rest too,
not just during a trip, so the evaluation period is not necessarily associated
with trip start/end.
1. We need to be careful about syncing since when duty cycling is on, we only
sync at the end of a trip. So for the power control, for example, we will never
really push anything. When duty cycling is turned off, we do push on an ongoing
basis, but if we transition from always on -> duty cycling, we may again miss
pushing some info.

Resulting modifications:
1. We set the settings when a new option is selected, in both cases. So the
toggle in the Calibration tab is gone.
1. In addition to documenting that the last evaluation performed should be AO,
    1. When transitioning from AO -> DC, force sync first
    1. Ensure that the power control is AO
This ensures that all data is pushed during the last AO -> defaults (DC) transition

Also added transitions - these are currently tied to UX clicks and don't depend
on the config setting being successful.
Also ensured that the evaluation card is hidden when the calibration is active