Change the eval settings to be more principled
This is to make them consistent with
https://github.com/e-mission/e-mission-eval/commit/cf5baffb2068160e90245a134229c20a53acb630

It also allows us to show which of the configurations being tested is active on
this phone and to store that information as part of the evaluation. Ideally,
this would go into a separate field, but the other messages for evaluation
transitions don't have this field, so for now, we munge the id to store this.

Let's see if we need to revisit this in the future.