Implement more complex user input matching
This fixes https://github.com/e-mission/e-mission-docs/issues/423
The new implementation works properly for trip end prompts and matches them
correctly to both processed and cleaned trips.

This is essentially a bunch of additional logging + the improved implementation
outlined in https://github.com/e-mission/e-mission-docs/issues/423

Since the implementation differs for processed v/s cleaned trips, the
`getUserInputForTrip` function now takes in the entire trip wrapper, not just
the properties. The changes to `list.js` are basically to pass in the right
kind of object.