Fix the formatting of the logs for cleaned trips
Just because the trips are cleaned does not mean that the inputs are.
Without this fix, some of the logs for cleaned trips (the ones comparing with
unprocessed inputs) would show undefined. While fixing that, also fix the other
messages to use the correct timezone. We may want to consider filling in the
`fmt_time` field for unprocessed trips and input messages to simplify this in
the future, but this works for now.

Testing done:

```
Input list = 2019-07-01T23:52:16-07:00(1562050336.88301) -> 2019-07-02T00:02:04-07:00(1562050924.886191) taxi logged at 1562051288.859764
2019-07-01T23:23:52-07:00(1562048632.865598) -> 2019-07-01T23:47:09-07:00(1562050029.868917) bike logged at 1562050048.594255
2019-06-19T11:08:34+05:30(1560922714.738) -> 2019-06-19T11:14:47+05:30(1560923087.839) Walk logged at 1560939263.386
2019-06-19T11:21:39+05:30(1560923499.1290061) -> 2019-06-19T11:33:33+05:30(1560924213.966) Rapido logged at 1560939289.797
2019-06-19T11:37:20+05:30(1560924440.0946546) -> 2019-06-19T11:44:37+05:30(1560924877.301) Ride in friends two-wheeler logged at 1560939294.646
2019-06-19T12:32:53+05:30(1560927773.399599) -> 2019-06-19T12:45:46+05:30(1560928546.527) Ride in friends two-wheeler logged at 1560939299.521
```