Fix the calibration transitions to be correct
Concretely:
- Move a bunch of the persistence code out to its own service
    - This also allows us to reduce the dependencies in each part
    - Use the new `storeBatteryReading` method along with proper ordering to
      ensure that all transitions are saved correctly and that we have battery
      readings in every case.

The proper ordering is:
- while starting a range (assuming tracking is turned off)
    1) mark range started.then(
        2) take a battery reading
        3) setConfig.then(turn on tracking if needed)
    )
we want to make sure that the battery reading happens *after* the range is
marked as started so that it will be retrieved properly when we read the values
within each range. But the battery reading and the set config can happen in parallel.
We can turn on the tracking after setting the config to avoid generating
additional transitions.

we don't really need to sync now although I guess we can if we really want to.
Leaving it off for now

- while ending a range (assuming tracking is turned on)
    (
    1) take a battery reading
    2) turn tracking off if needed.then(set config)
    ).then(end the range).then(sync)

Couple of notes, we don't really need the setConfig here since we are turning
tracking off completely anyway. However, if we ensure that the tracking off
config has duty cycling turned off, the sync will push all entries immediately.
woo!  And again, we want to ensure that the battery reading is *before* the
range so that it gets picked up properly in the range. Note that with this
pattern, we will have two configs for each range, so we may need to change the
validation. Although the first one should still match the spec for the range,
so we won't need to change it too much.

- while changing a range (tracking is on, just changing the config)
    1) take a battery reading
        .then(end the old range)
        .then(start the new range)
        .then(take a battery reading)
        .then(turn tracking off if needed based on the new config)
        .then(set config)

    This should ensure that we:
        - have a battery reading before ending the old range
        - have a battery reading at the beginning of the new range
        - set the new config after turning tracking off if necessary so that we
          won't turn off/on unnecessarily