Fix regression for trip end
Fix regression caused by reordering in 706c57b4947ed2251b166f1141f97e6e17bdc3b6
Since we moved the transition generation to the end, the section id was blanked
out and we got an undefined error.

We now store the initial trip **and** section id and use the to generate the
transition.

Testing done:
- Went through a local spec and ensured that the pushed values had the stop
  trip transition and the final stop section transition