Bump up all plugins to the latest version
Verified via `vimdiff` from the covid19 app.
Alas, we can't use cherry-pick because of the unrelated changes (e.g. app name, app version)

e-mission also uses plugins that are not in the covid19 app, so we upgrade
those to the most recent versions as well.

A particularly important update is the webview update to 3.2.0 to handle the
WKWebView issue.

We also update the push plugin to the most recent version. This has been flaky
in the past, but we were already at v2, so hopefully this is not a big deal.
The push plugin also requires `google-services.json` and
`GoogleService-Info.plist` to be downloaded from the Firebase console.

These are obviously different for each different native build, AND
we are obviously not going to check in the versions for the base app.

So we check in fake versions of the files which have the same format as the
original files but with faked out ids and project names and modify the setup
scripts to copy them over.

This will work for the build CI for now - for actual testing CI, we will need
to set up encrypted storage as part of the workflow. Which will only run on
pushes to the main repo not on the forks, so it will be complicated to test.