Reformat the actions to be consistent with the new spec
Note that this is consistent with the sample app, but is NOT consistent with
the master branch. There is a regression in the master branch - inline actions
don't work although they are in the README. Specifying the action group ahead
of time does work in isolation, but seems to require configuring the action
group just before the schedule.
https://github.com/e-mission/e-mission-transition-notify/issues/25#issuecomment-660820728

So we go back to `0.9.0-beta.3` and the working spec