Remove references to ngCordova
ngCordova is now deprecated
https://github.com/ionic-team/ng-cordova#readme

We could move to IonicNative, but not sure how much longer we are going to
stick with Ionic.

Going back to regular cordova plugin calls, which we use in other places as well.
The only places where we were using ngCordova wrappers was for the in-app
browser and the email composer

+ changed the options for the in-app browser to match the new spec in which the
options are in a single string instead of a JSON dict. Without this change, the
browser is not launched properly