Tweak the CI build further
The CI build was not failing although there were errors.
That is because we switched from `source` to `bash`, so the job doesn't crash any more.
Instead it exits.

Moving the checks to the next action so that the setup is the last action and
exiting with a failure code causes the action to fail.

+ try to add some other commands to the cocoapods uninstall